requirements.txt
pandas>=1.5.0
python-dateutil>=2.8.0
pyyaml>=6.0
pycoingecko>=3.1.0
pytest>=7.0.0
fpdf2>=2.7.0
openpyxl>=3.1.0
argparse
requests>=2.28.0
flask>=2.3.0
numpy>=1.21.0
matplotlib>=3.5.0
seaborn>=0.11.0
plotly>=5.0.0
dash>=2.0.0
dash-bootstrap-components>=1.0.0